+++ common/src/main/java/org/endeavourhealth/recordviewer/common/models/ReferralRequestFull.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class ReferralRequestFull {
 
     private String id;
-    private int patientId;
+    private long patientId;
     private String practitionerId;
     private String recipientOrganizationId;
     private String intent;
@@ -25,11 +25,11 @@ public class ReferralRequestFull {
         return this;
     }
 
-    public int getPatientId() {
+    public long getPatientId() {
         return patientId;
     }
 
-    public ReferralRequestFull setPatientId(int patientId) {
+    public ReferralRequestFull setPatientId(long patientId) {
         this.patientId = patientId;
         return this;
     }
